backend/pyproject.toml
[tool.poetry]
name = "algo-trading-api"
version = "0.1.0"
description = "FastAPI backend for algo trading web application"
authors = ["Trading Team <team@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.4"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.6.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"
aiofiles = "^24.1.0"
python-socketio = "^5.11.4"
redis = "^5.2.0"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"
httpx = "^0.27.2"
alpaca-py = "^0.28.0"
tenacity = "^8.2.3"
pandas = "^2.2.0"
numpy = "^2.0.0"
pandas-ta = "0.4.71b0"
apscheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
mypy = "^1.13.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
